# openapi/openapi.yml
openapi: 3.1.0
info:
  title: Roe AI API
  version: 1.0.0
  description: Complete API documentation for Roe AI platform
paths:
...
  /v1/datasets/{dataset_id}/files/{id}/:
    get:
      operationId: v1_datasets_files_retrieve
      description: "Mixin for organization access validation.\n\nAutomatically resolves\
        \ and validates organization access on every request,\nmaking the organization\
        \ available as self.organization.\n\nFor API key auth:  organization is derived\
        \ from the key (no org_id param needed).\nFor user auth:     organization_id\
        \ must be provided in the request\n                   (kwargs / headers /\
        \ body / query params).\n\nUsage:\n    class MyView(BaseOrganizationAccessMixin,\
        \ APIView):\n        def get(self, request):\n            # self.organization\
        \ is automatically available\n            return Response({\"org\": self.organization.name})"
      parameters:
      - in: path
        name: dataset_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: organization_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Organization ID. This is required for access control. It can
          be provided via query or request body depending on the endpoint.
      tags:
      - v1
      - sdk
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
          description: ''
    put:
      operationId: v1_datasets_files_update
      description: "Mixin for organization access validation.\n\nAutomatically resolves\
        \ and validates organization access on every request,\nmaking the organization\
        \ available as self.organization.\n\nFor API key auth:  organization is derived\
        \ from the key (no org_id param needed).\nFor user auth:     organization_id\
        \ must be provided in the request\n                   (kwargs / headers /\
        \ body / query params).\n\nUsage:\n    class MyView(BaseOrganizationAccessMixin,\
        \ APIView):\n        def get(self, request):\n            # self.organization\
        \ is automatically available\n            return Response({\"org\": self.organization.name})"
      parameters:
      - in: path
        name: dataset_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: organization_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Organization ID. This is required for access control. It can
          be provided via query or request body depending on the endpoint.
      tags:
      - v1
      - sdk
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FileRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
          description: ''
    patch:
      operationId: v1_datasets_files_partial_update
      description: "Mixin for organization access validation.\n\nAutomatically resolves\
        \ and validates organization access on every request,\nmaking the organization\
        \ available as self.organization.\n\nFor API key auth:  organization is derived\
        \ from the key (no org_id param needed).\nFor user auth:     organization_id\
        \ must be provided in the request\n                   (kwargs / headers /\
        \ body / query params).\n\nUsage:\n    class MyView(BaseOrganizationAccessMixin,\
        \ APIView):\n        def get(self, request):\n            # self.organization\
        \ is automatically available\n            return Response({\"org\": self.organization.name})"
      parameters:
      - in: path
        name: dataset_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: organization_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Organization ID. This is required for access control. It can
          be provided via query or request body depending on the endpoint.
      tags:
      - v1
      - sdk
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedFileRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedFileRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedFileRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
          description: ''
    delete:
      operationId: v1_datasets_files_destroy
      description: "Mixin for organization access validation.\n\nAutomatically resolves\
        \ and validates organization access on every request,\nmaking the organization\
        \ available as self.organization.\n\nFor API key auth:  organization is derived\
        \ from the key (no org_id param needed).\nFor user auth:     organization_id\
        \ must be provided in the request\n                   (kwargs / headers /\
        \ body / query params).\n\nUsage:\n    class MyView(BaseOrganizationAccessMixin,\
        \ APIView):\n        def get(self, request):\n            # self.organization\
        \ is automatically available\n            return Response({\"org\": self.organization.name})"
      parameters:
      - in: path
        name: dataset_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - name: organization_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Organization ID. This is required for access control. It can
          be provided via query or request body depending on the endpoint.
      tags:
      - v1
      - sdk
      responses:
        '204':
          description: No response body
...
  /v1/webhooks/:
    get:
      operationId: v1_webhooks_list
      description: "Mixin for organization access validation.\n\nAutomatically resolves\
        \ and validates organization access on every request,\nmaking the organization\
        \ available as self.organization.\n\nFor API key auth:  organization is derived\
        \ from the key (no org_id param needed).\nFor user auth:     organization_id\
        \ must be provided in the request\n                   (kwargs / headers /\
        \ body / query params).\n\nUsage:\n    class MyView(BaseOrganizationAccessMixin,\
        \ APIView):\n        def get(self, request):\n            # self.organization\
        \ is automatically available\n            return Response({\"org\": self.organization.name})"
      parameters:
      - in: query
        name: organization_id
        schema:
          type: string
          format: uuid
        description: Organization ID. Required for JWT auth; inferred from API key
          when using an Organization API Key.
      tags:
      - v1
      - sdk
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
          description: ''
    post:
      operationId: v1_webhooks_create
      description: "Mixin for organization access validation.\n\nAutomatically resolves\
        \ and validates organization access on every request,\nmaking the organization\
        \ available as self.organization.\n\nFor API key auth:  organization is derived\
        \ from the key (no org_id param needed).\nFor user auth:     organization_id\
        \ must be provided in the request\n                   (kwargs / headers /\
        \ body / query params).\n\nUsage:\n    class MyView(BaseOrganizationAccessMixin,\
        \ APIView):\n        def get(self, request):\n            # self.organization\
        \ is automatically available\n            return Response({\"org\": self.organization.name})"
      tags:
      - v1
      - sdk
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWebhook'
          description: ''
      parameters:
      - name: organization_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
        description: Organization ID. This is required for access control. It can
          be provided via query or request body depending on the endpoint.
...
components:
  schemas:
...
    ContentTypeEnum:
      enum:
      - application/pdf
      - application/json
      - audio/mpeg
      - audio/mp3
      - audio/wav
      - audio/x-m4a
      - video/mp4
      - image/jpeg
      - image/png
      - text/plain
      - text/html
      - text/xml
      - text/csv
      - text/markdown
      - audio/*
      - video/*
      - image/*
      - text/*
      - '*/*'
      type: string
      description: '* `application/pdf` - APPLICATION_PDF

        * `application/json` - APPLICATION_JSON

        * `audio/mpeg` - AUDIO_MPEG

        * `audio/mp3` - AUDIO_MP3

        * `audio/wav` - AUDIO_WAV

        * `audio/x-m4a` - AUDIO_M4A

        * `video/mp4` - VIDEO_MP4

        * `image/jpeg` - IMAGE_JPEG

        * `image/png` - IMAGE_PNG

        * `text/plain` - TEXT_PLAIN

        * `text/html` - TEXT_HTML

        * `text/xml` - TEXT_XML

        * `text/csv` - TEXT_CSV

        * `text/markdown` - TEXT_MARKDOWN

        * `audio/*` - AUDIO

        * `video/*` - VIDEO

        * `image/*` - IMAGE

        * `text/*` - TEXT

        * `*/*` - ANY'
...
    CreateWebhook:
      type: object
      description: Serializer for creating webhooks
      properties:
        name:
          type: string
          description: User-friendly name for this webhook
          maxLength: 255
        url:
          type: string
          format: uri
          description: Webhook endpoint URL
          maxLength: 500
        secret:
          type: string
          maxLength: 255
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom headers to include in webhook requests (key-value pairs)
        events:
          type: array
          items:
            type: string
          description: List of events to trigger this webhook (e.g., ['triggered'])
        is_active:
          type: boolean
        alerts:
          type: array
          items:
            type: string
            format: uuid
      required:
      - name
      - url
    CreateWebhookRequest:
      type: object
      description: Serializer for creating webhooks
      properties:
        name:
          type: string
          minLength: 1
          description: User-friendly name for this webhook
          maxLength: 255
        url:
          type: string
          format: uri
          minLength: 1
          description: Webhook endpoint URL
          maxLength: 500
        secret:
          type: string
          maxLength: 255
        headers:
          type: object
          additionalProperties:
            type: string
            minLength: 1
          description: Custom headers to include in webhook requests (key-value pairs)
        events:
          type: array
          items:
            type: string
            minLength: 1
          description: List of events to trigger this webhook (e.g., ['triggered'])
        is_active:
          type: boolean
        alerts:
          type: array
          items:
            type: string
            format: uuid
      required:
      - name
      - url
...
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        dataset:
          type:
          - string
          - 'null'
          format: uuid
        hash:
          type: string
          maxLength: 1024
        name:
          type: string
          maxLength: 1024
        size:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        content_type:
          $ref: '#/components/schemas/ContentTypeEnum'
        blob_id:
          type: string
          maxLength: 1024
        creator:
          allOf:
          - $ref: '#/components/schemas/UserInfo'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        meta: {}
        metadata:
          type: object
          additionalProperties:
            type: string
          readOnly: true
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        deleted_by:
          allOf:
          - $ref: '#/components/schemas/UserInfo'
          readOnly: true
      required:
      - blob_id
      - content_type
      - created_at
      - creator
      - deleted_at
      - deleted_by
      - hash
      - id
      - metadata
      - name
      - size
...
    FileRequest:
      type: object
      properties:
        dataset:
          type:
          - string
          - 'null'
          format: uuid
        hash:
          type: string
          minLength: 1
          maxLength: 1024
        name:
          type: string
          minLength: 1
          maxLength: 1024
        size:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        content_type:
          $ref: '#/components/schemas/ContentTypeEnum'
        blob_id:
          type: string
          minLength: 1
          maxLength: 1024
        meta: {}
      required:
      - blob_id
      - content_type
      - hash
      - name
      - size
...
    PatchedFileRequest:
      type: object
      properties:
        dataset:
          type:
          - string
          - 'null'
          format: uuid
        hash:
          type: string
          minLength: 1
          maxLength: 1024
        name:
          type: string
          minLength: 1
          maxLength: 1024
        size:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
        content_type:
          $ref: '#/components/schemas/ContentTypeEnum'
        blob_id:
          type: string
          minLength: 1
          maxLength: 1024
        meta: {}
...
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        display_name:
          type: string
          readOnly: true
      required:
      - display_name
      - email
      - id
...
    Webhook:
      type: object
      description: Serializer for Webhook model
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          description: User-friendly name for this webhook
          maxLength: 255
        url:
          type: string
          format: uri
          description: Webhook endpoint URL
          maxLength: 500
        secret:
          type: string
          maxLength: 255
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom headers to include in webhook requests (key-value pairs)
        events:
          type: array
          items:
            type: string
          description: List of events to trigger this webhook (e.g., ['triggered'])
        is_active:
          type: boolean
        alerts:
          type: array
          items:
            type: string
          description: Alert IDs
          readOnly: true
        last_triggered_at:
          type: string
          format: date-time
        last_success_at:
          type: string
          format: date-time
        last_failure_at:
          type: string
          format: date-time
        failure_count:
          type: integer
          readOnly: true
          description: Number of consecutive failures
        created_by:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - alerts
      - created_at
      - created_by
      - failure_count
      - id
      - name
      - updated_at
      - url
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: https://api.roe-ai.com
  description: Production API
...